Python(chore): Reorganize pytest code (#610)
